Add things about color generation
XEP-0392 is about to lose this information [1] because it is out of
scope for the XSF and the suggestions in XEP-0392 have been too
restrictive.

This addition here provides the information which is about to be
removed from XEP-0392, as well as valuable feedback from
@mar-v-in about contrast ratio considerations. Restrictions about
the form of dummy avatars are dropped intentionally and entirely,
because those may easily vary across platforms, as long as the
color is prominently present.

   [1]: https://github.com/xsf/xeps/pull/845